+++ example_glium/src/main.rs
@@ -12,7 +12,7 @@ struct MyApp {
 
 impl egui::app::App for MyApp {
     /// Called each time the UI needs repainting, which may be many times per second.
-    /// Put your widgets into a `SidePanel`, `TopPanel`, `Window` or into `ctx.background_ui()`.
+    /// Put your widgets into a `SidePanel`, `TopPanel`, `CentralPanel`, `Window` or `Area`.
     fn ui(
         &mut self,
         ctx: &std::sync::Arc<egui::Context>,